Patch from Jerret Taylor to fix caching issue with controller and action names
Original message:

I noticed a bug with the caching - if you caches_embedded an action, the way it registers that as being cachable will cause it to cache that action for any controller during the request - and seems to leak into other requests as well...An example, a controller calling two embedded actions on two seperate controllers, both with the :show action - if one is set to cache_embedded and the other is not, both will.

I wrote some tests and a patch to fix it - I also added a test 'test_should_not_return_from_cache_if_params_are_different' just to make sure, which passed right away without any code changes, but i figured I'd leave it anyway, more tests is always a good thing!


git-svn-id: https://dev.notso.net/svn/rails/plugins/embedded_actions/trunk@149 71bf3dff-0d17-0410-8985-d5c41cf5d41c